refactor: commit 컴포넌트 분리 및 메소드 분리(#18)
